add middlewares to route and edit user ban controller

+++ controllers/user.js
@@ -2,13 +2,16 @@ const userModel = require('./../models/user');
 const banModel = require('./../models/ban-phone');
 
 exports.banUser = async (req, res) => {
-    const mainUser = await userModel.findOne({_id: req.params.id}).lean();
+    const mainUser = await userModel.findOne({ _id: req.params.id }).lean();
+    if (!mainUser) {
+        return res.status(204).json({ message: 'user has banned and deleted from DB' })
+    }
     const banUserResult = await banModel.create({
         phone: mainUser.phone
     });
     if (banUserResult) {
-        await userModel.deleteOne({_id: mainUser._id})
-        return res.status(200).json({message: 'user banned successfully'})
+        await userModel.deleteOne({ _id: mainUser._id })
+        return res.status(200).json({ message: 'user banned successfully' })
     }
-    return res.status(500).json({message: 'server error in ban user'})
+    return res.status(500).json({ message: 'server error in ban user' })
 }
